Rename puzzlesection file to be pascal case
