+++ src/views/Lessons/PuzzleSections.ts
